requirements_ci.txt
pandas-stubs
pyinstaller
types-requests
mypy
flake8
pytest
